Add tracker support for Piwik
While we do not use Google Analytics on the RMS page to respect his
privacy preferences, we do use Piwik for this page. Piwik is free/libre
software and we do not share it's result with other companies. It's also
configured to respect the users privacy by anonymizing and watching the
'do not track' setting.